Fixed problems with the jetty-plugin.
Now you can execute the application as exploded from mpango project or choose to run as you like from the mpango-web project (both should run the application on http://localhost:8080/mpango).

Important to note: 
In case of running the application from mpango project, make sure you before run a "mvn package" so that everything is packaged and ready to be serverd.
In case of running the application from mpango-web, you have to make sure to have an updated mpango-core version installed in your Maven repository (this is done by executing mvn install).


git-svn-id: https://svn.code.sf.net/p/mpango/code/trunk@84 adca2377-96ed-487d-85b0-23fc967ffdab